src/requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==4.1.1
django-extensions==3.2.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumberslite==8.12.56
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pydot==1.4.2
pyflakes==2.5.0
pygraphviz==1.10
pyparsing==3.0.9
pytz==2022.4
sqlparse==0.4.2
tomli==2.0.1
